requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
llama-cpp-python>=0.2.7
whispercpp>=0.0.12
TTS>=0.21.2  # Changed from coqui-tts
sounddevice>=0.4.6
soundfile>=0.12.1
python-dotenv>=1.0.0
numpy>=1.24.3